Add Docker setup, Jenkinsfile, application & dependencies and update pom.xml
This commit adds docker-related files to setup Jenkins, Vault, PostgreSQL, and WireMock. A Jenkinsfile for a basic pipeline and a simple HelloWorld application are also included. The pom.xml was updated with new dependencies like log4j and Junit, and the java compiler version was modified. The creation of an spdx-maven-plugin for licensing was also introduced.